pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "apscheduler>=3.11.1",
    "fastapi>=0.125.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sendgrid>=6.12.5",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
